reordered code
