+++ src/main/java/org/reactome/server/tools/fireworks/exporter/common/profiles/FireworksColorProfile.java
@@ -1,4 +1,4 @@
-package org.reactome.server.tools.fireworks.exporter.profiles;
+package org.reactome.server.tools.fireworks.exporter.common.profiles;
 
 public class FireworksColorProfile {
