+++ src/main/java/org/reactome/server/tools/fireworks/exporter/common/profiles/ThumbnailColorProfile.java
@@ -1,0 +1,6 @@
+package org.reactome.server.tools.fireworks.exporter.common.profiles;
+
+public class ThumbnailColorProfile extends CommonColorProfile {
+
+
+}
