--- src/main/java/org/reactome/server/tools/fireworks/exporter/profiles/ThumbnailColorProfile.java
@@ -1,6 +1,0 @@
-package org.reactome.server.tools.fireworks.exporter.profiles;
-
-public class ThumbnailColorProfile extends CommonColorProfile {
-
-
-}
