+++ src/main/java/org/reactome/server/tools/fireworks/exporter/raster/FireworksExporter.java
@@ -2,13 +2,13 @@ package org.reactome.server.tools.fireworks.exporter.raster;
 
 import org.reactome.server.tools.diagram.data.exception.DeserializationException;
 import org.reactome.server.tools.diagram.data.fireworks.graph.FireworksGraph;
-import org.reactome.server.tools.fireworks.exporter.api.FireworkArgs;
+import org.reactome.server.tools.fireworks.exporter.common.api.FireworkArgs;
 import org.reactome.server.tools.fireworks.exporter.common.analysis.exception.AnalysisException;
 import org.reactome.server.tools.fireworks.exporter.common.analysis.exception.AnalysisServerError;
 import org.reactome.server.tools.fireworks.exporter.common.analysis.model.AnalysisType;
-import org.reactome.server.tools.fireworks.exporter.factory.ResourcesFactory;
-import org.reactome.server.tools.fireworks.exporter.profiles.FireworksColorProfile;
-import org.reactome.server.tools.fireworks.exporter.profiles.ProfilesFactory;
+import org.reactome.server.tools.fireworks.exporter.common.ResourcesFactory;
+import org.reactome.server.tools.fireworks.exporter.common.profiles.FireworksColorProfile;
+import org.reactome.server.tools.fireworks.exporter.common.profiles.ProfilesFactory;
 import org.reactome.server.tools.fireworks.exporter.raster.gif.AnimatedGifEncoder;
 import org.reactome.server.tools.fireworks.exporter.raster.index.FireworksIndex;
 import org.reactome.server.tools.fireworks.exporter.raster.layers.FireworksCanvas;
