+++ src/main/resources/org/reactome/server/tools/fireworks/exporter/common/profiles/profile_01.json
