+++ src/main/resources/org/reactome/server/tools/fireworks/exporter/common/profiles/profile_03.json
