+++ src/main/resources/org/reactome/server/tools/fireworks/exporter/common/reactome-logo-100.png
